use dataSpace elements when defaulting to first class
